requirements.txt
ipywidgets
openpyxl
pandas
pandas_market_calendars
plotly
python-dotenv
requests_html
scikit-learn
selenium
statsmodels
yahoo-fin
yfinance
